Add tests to TestAnimal.
Add error handling to Animal. In a real app this error handling would be done at time of data entry (by the user). But I'm not planning a main method, so its in the class itself.

The error handling in Animal can be massively simplified by making a method for each property and calling the same method at instantiation and in the respective setter. But I have to make some mistakes to demonstrate I know better, and not that I'm lucking into optimised code lol.